refactor: Refactor and remove redundant code in fuzzer mutation logic. Move MakeMessage to `bit_mutations.rs`.
